@@ -10,16 +10,23 @@
         "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --no-progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js"
     },
     "devDependencies": {
-        "axios": "^0.18",
-        "bootstrap": "^4.1.0",
-        "cross-env": "^5.1",
-        "jquery": "^3.2",
-        "laravel-mix": "^4.0.7",
-        "lodash": "^4.17.5",
-        "popper.js": "^1.12",
+        "cross-env": "^5.2.0",
+        "css-loader": "^2.1.1",
+        "laravel-mix": "^4.0.16",
+        "lodash": "^4.17.11",
+        "node-sass": "^4.12.0",
         "resolve-url-loader": "^2.3.1",
-        "sass": "^1.15.2",
         "sass-loader": "^7.1.0",
-        "vue": "^2.5.17"
+        "vue-loader": "^15.7.0",
+        "vue-template-compiler": "^2.6.10"
+    },
+    "dependencies": {
+        "axios": "^0.19.0",
+        "js-cookie": "^2.2.0",
+        "sweetalert2": "^8.11.7",
+        "vue": "^2.6.10",
+        "vue-avatar": "^2.1.7",
+        "vue-router": "^3.0.6",
+        "vuex": "^3.1.1"
     }
 }
