@@ -1,32 +1,28 @@
-/**
- * First we will load all of this project's JavaScript dependencies which
- * includes Vue and other libraries. It is a great starting point when
- * building robust, powerful web applications using Vue and Laravel.
- */
+import Vue from 'vue'
 
-require('./bootstrap');
+import App from '~/Layouts/App'
+import router from '~/Router'
+import store from '~/Store'
+import mixin from "./mixins"
 
-window.Vue = require('vue');
+import '~/Plugin'
+import '~/components'
+import axios from "axios";
+import swal from "sweetalert2"
 
-/**
- * The following block of code may be used to automatically register your
- * Vue components. It will recursively scan this directory for the Vue
- * components and automatically register them with their "basename".
- *
- * Eg. ./components/ExampleComponent.vue -> <example-component></example-component>
- */
+// window.$ = window.jQuery = require('jquery');
 
-// const files = require.context('./', true, /\.vue$/i);
-// files.keys().map(key => Vue.component(key.split('/').pop().split('.')[0], files(key).default));
+store.$axios=axios
+Vue.prototype.$axios=axios
+Vue.prototype.$swal=swal
 
-Vue.component('example-component', require('./components/ExampleComponent.vue').default);
+Vue.prototype.$_env=window.config
+delete window["config"]
 
-/**
- * Next, we will create a fresh Vue application instance and attach it to
- * the page. Then, you may begin adding components to this application
- * or customize the JavaScript scaffolding to fit your unique needs.
- */
+const vue=new Vue({
+    mixin,
+    ...App,
+    router,
+    store,
+}).$mount("#app")
 
-const app = new Vue({
-    el: '#app',
-});
